[Fix] : ADBackground 스크롤시 refresh 현상 수정
Git : #14
