+++ Projects/UI/AD_UI/Sources/MakeADViews/Step1_UploadADrawingView/UploadADrawingView.swift
@@ -44,7 +44,6 @@ struct UploadADrawingView: ADUI {
           Spacer()
         }
         .padding()
-        .background(ADBackground())
       }
     }
   }
