+++ Projects/UI/AD_UI/Sources/MakeADViews/Step2_FindingTheCharacterView/FindingTheCharacterView.swift
@@ -41,7 +41,6 @@ struct FindingTheCharacterView: ADUI {
           Spacer()
         }
         .padding()
-        .background(ADBackground())
       }
     }
   }
