+++ Projects/UI/AD_UI/Sources/RootView.swift
@@ -14,38 +14,10 @@ public struct RootView: View {
   public init() {}
   
   public var body: some View {
-    GeometryReader { proxy in
-      Group {
-        if !isTapStartButton {
-          OnBoardingView(isTapStartButton: $isTapStartButton)
-        } else {
-          MakeADView()
-        }
-      }
-      .environment(\.mainWindowSize, proxy.size)
-      .environment(\.mainWindowRect, .init(origin: .init(), size: proxy.size))
+    if !isTapStartButton {
+      OnBoardingView(isTapStartButton: $isTapStartButton)
+    } else {
+      MakeADView()
     }
-  }
-}
-
-/// Use Like
-/// @Environment(\.mainWindowSize) var mainWindowSize
-private struct MainWindowSizeKey: EnvironmentKey {
-  static let defaultValue: CGSize = .zero
-}
-
-private struct MainWindowRectKey: EnvironmentKey {
-  static let defaultValue: CGRect = .zero
-}
-
-extension EnvironmentValues {
-  var mainWindowSize: CGSize {
-    get { self[MainWindowSizeKey.self] }
-    set { self[MainWindowSizeKey.self] = newValue }
-  }
-  
-  var mainWindowRect: CGRect {
-    get { self[MainWindowRectKey.self] }
-    set { self[MainWindowRectKey.self] = newValue }
   }
 }
